feat(pic16f87xa-hal): working XC8 v3.10 target build (.hex for all 4 devices)
The MPLAB X / XC8 Makefile previously targeted v2.40 and never linked:
under xc8-cc v3.x the compile emits .p1 p-code (not .o), the link needs
-ginhx32 (not -ohex:), and v3.x requires -mdfp for device support.

- Makefile: rework for xc8-cc v3.x — -mdfp (overridable DFP_DIR),
  .p1 intermediates, single -ginhx32 link step, order-only build/ prereq.
- pic16f87xa.h: cast SFR addresses through uintptr_t so XC8 stops warning
  about uint8_t -> pointer conversion.
- pic16f87xa_gpio.c: revert tris_addr/port_addr/port_width to switch
  functions. The macro form referenced GPIOD/GPIOE lexically, which are
  undeclared on 28-pin parts (873A/876A) — && short-circuits at runtime,
  not compile time, so those builds broke. The real XC8 fix is dropping
  const from the function-initialized locals (XC8 rejects const init from
  a function call); the "constant initializer" rationale for the macros
  did not apply, as the helpers are only called at runtime.
- example_blink.c: make dual-mode. It called pic16f87xa_sim_* unconditionally
  and so could never link on a target. Sim path (host CMake) is unchanged;
  the new target path polls Timer0 to blink RB0 and refreshes the WDT.
- mcu README: v2.40 -> v3.10.

Verified: host sim build + all 11 example tests pass; all four devices
(873A/874A/876A/877A) produce a real <MCU>-firmware.hex with zero errors.

Known gap: the HAL still has no XC8 interrupt-vector dispatcher, so on
real silicon interrupts cannot reach the weak handlers; the target blink
polls instead.

+++ pic16f87xa-hal/include/pic16f87xa.h
@@ -158,11 +158,14 @@ typedef enum {
   #define pic16f87xa_sfr_write8(addr, v)                                      \
       do { pic16f87xa_sim_sfr[(uint16_t)(addr)] = (uint8_t)(v); } while (0)
 #else
-  /* Real target — direct volatile access. */
-  #define PIC16F87XA_SFR_PTR(addr)     ((volatile uint8_t *)(addr))
-  #define pic16f87xa_sfr_read8(addr)   (*(volatile uint8_t *)(addr))
+  /* Real target — direct volatile access.  Cast the SFR address
+   * through `uintptr_t` so XC8 does not warn about converting a
+   * `uint8_t` to a pointer. */
+  #include <stdint.h>
+  #define PIC16F87XA_SFR_PTR(addr)     ((volatile uint8_t *)(uintptr_t)(addr))
+  #define pic16f87xa_sfr_read8(addr)   (*(volatile uint8_t *)(uintptr_t)(addr))
   #define pic16f87xa_sfr_write8(addr, v)                                      \
-      do { *(volatile uint8_t *)(addr) = (uint8_t)(v); } while (0)
+      do { *(volatile uint8_t *)(uintptr_t)(addr) = (uint8_t)(v); } while (0)
 #endif
 /** @} */
 
@@ -170,7 +173,7 @@ typedef enum {
 #if defined(PIC16F87XA_USE_SIMULATOR)
   #define PIC16F87XA_REG8(addr)  (pic16f87xa_sim_sfr[(uint16_t)(addr)])
 #else
-  #define PIC16F87XA_REG8(addr)  (*(volatile uint8_t *)(addr))
+  #define PIC16F87XA_REG8(addr)  (*(volatile uint8_t *)(uintptr_t)(addr))
 #endif
 
 #ifdef __cplusplus
